requirements.txt
av==13.1.0
decord==0.6.0
diffusers==0.31.0
einops==0.8.0
imageio==2.27.0
matplotlib==3.9.2
moviepy==1.0.3
numpy==2.1.3
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.0.0
Requests==2.32.3
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.5
transformers==4.46.0
